webui: tests: move cockpit-ws spawning to the machine_install script instead of the ks file
machine.wait_boot() is now redundant in the test/webui_testvm.py as it's
already checked in the test/machine_install.py.

Redundant is as well the curling of the webui URL in the
test/machine_install.py script. Since now the spawning of the
cockpit-ws is done after the VM creation, through SSH, checking the
accessibility of the webui is not required.