webui: rewrite language selection component to a class component
Usually I prefer using functional components, as they require less
boilerblate code, thus increasing the code reability.
However, when complicated state management needs to be done in the
inialization (in the component mounting), the usage of useEffect hook can get quite
complicated.

Let's convert the LanguageSelector component to be a class component for
this reason, and use the more straightforward componentDidMount
lifecycle method for doing the initialization.

Lastly, this fixes one bug, and thus clicking on the [x] button on the
typeahead select previously did not clear the state.