webui: introduce some logic for which steps the users can access
The user selected configuration is currently only applied when the user
clicks 'Next' in a tab. For this reason, we need to make sure that the
user has submitted all of the previous steps before accessing a
new one.

Otherwise it's probable that we will end up in the `review step`, without having
completed any storage configuration tasks ran.

The language step is special, in the meaning that once submitted it
can't be revisited. (This is actually prompt to change according to designs)