webui: start using custom webui-desktop script instead of cockpit-desktop
webui-desktop is a patched version of cockpit-desktopi [1], with the
slight difference that it does not call the unshare command.

The reason for this change is that the /bin/unshare is not present in
rawhide's  boot.iso, as it get's filtered out by lorax [0]:

Ideally, we need to teach lorax to not filter out this command, and keep
using the original cockpit-desktop script [1]:

The approach to add linux-utils-core package with the updates image,
brings the /bin/mount and other binaries permissions into a bad state,
making the installation to fail at an early stage.

[0]: https://github.com/weldr/lorax/blob/master/share/templates.d/99-generic/runtime-cleanup.tmpl#L294
[1]: https://github.com/cockpit-project/cockpit/blob/main/src/ws/cockpit-desktop.in