webui: add support for re-scanning disks
The UI now offerts the `re-scan` button for re-fetching disks in the
storage step.

For testing this feature, the test VM creation script had to be
extended.
In specific we added the `add_disk` method to the machine_install script.

Machine.add_disk() is a method specific to our custom VM creation scipt and
not part of MachineCase, and thus is not available for @nondestructive tests.

This commit additionaly fixes displaying more than one disks which was
previously broken generally.

Note: This commit can be used as a reference to also add other device types on
request to the test VM.