webui: tests: Wait for UI initialization in Installer.open()
Tests like `TestBasic.testNavigation` start with a `wait_not_present()`,
which is a race condition: This will always be true right after opening
the URL, as the page gets loaded asynchronously. To fix this
generically, let Installer.open() wait for the page to be sufficiently
loaded to show the steps menu and the welcome screen.