webui: Use ProgressStepper on the installation progress screen
Use the PatternFly ProgressStepper component instead of a progress bar
on the installation progress screen.

Also drop the "Installation progress" header as it does not really
fit in to the way how the progress screen with progress stepper looks
like.

NOTE: The progress stepping is currently hardcoded and based on MVP
order of installation tasks. This done due to DBus API for installation
phases not yet being available and to gather quick feedback for using
ProgressStepper on the progress page. If we keep ProgressStepper it is
expected that a regular DBus API will be added for both installation
phase discovery and phase transition signal.