webui: Wait longer for installation to fail
With the tar payload we use for testing the installation is expected to
fail in a specific manner.

While this should be deterministic, it can take a different amount of
time each run, based on how loaded is the testing infrastructure.

So to avoid unnecessary flakes bump the timeout for this one call to
300 seconds, giving hopefully ample time for the error notification
we are looking for to show up, even when the test infra is under high
load.