webui: Disable "Next" button if no disks are selected
Disable the "Next" button in the Wizard if no disks are selected
in the installation destination step.

This is using the isFormValid property, which is already used by
the language selection step.

But in this case as the installation destination step is not the first one,
we need to ale take care to reset form validation status back to default
when the step is exited.

The idea is that we always set it to default (true) in the functions
that govern step switching & each step can then do the validation
in their refresh handler (usually the useEffect() call).

Also add a tooltip so that users know what to do to activate the next
button.

As this also changes the UI workflow, adjust the unit tests accordingly.