webui: clean up prepare-updates-img
Tweak the script to make it easier to read and change.  Also: deal with
changing cockpit versions more elegantly: we now download the new
versions, even if the old ones are already on disk, and we also avoid
wildcards to ensure that we never try to install any old ones that are
lying around.

Add some .gitignore for the tmp directory we use for RPMs.