WebUI tweak local test execution
This change improves the current state of how to run locally the cockpit
tests.

This change includes:
- Do not force people to call `make rpms`
This is not required if we build the srpm files as part of the
prepare-updates-img script which creates a new updates-image for
testing.

- Use `pushd` and `popd` in the prepare-updates-img
This makes the script a bit more robust

- Do not force people to call `make bots` and `make test/common`
separately.
Could be easily done by one call.

- Improve documentation
The Readme.rst file reflects the new solution. Should be simpler and
don't have unnecessary steps.
Add explicit note how to work on Fedora SilverBlue.