WebUI: use StorageScenarioId in all components
Instead of relying on the selected initialization method for selecting
the storage configuration and review screen use the Wizard's
storageScenarioId. This allows us to later add a new storage
configuration which can have the same initialization as another
scenario. It also makes it a lot clearer what storage scenario was
chosen in the code as it no longer relies on the initialization method.