webui: start using useReducer for managing complicated state objects centrally
useReducer is a hook that allows to extract state logic outside of the
components into a state store. [1] [2]

Anaconda is consuming quite complicated objects from the anaconda dbus
APIs. The fetched API data will soon need to be reused across components
while the app grows and get's more complex.

Extracting part of the state outside of the components makes API data
manipulation and fetching a bit cleaner and easier to reuse.

[1] https://react.dev/reference/react/useReducer
[2] https://react.dev/learn/extracting-state-logic-into-a-reducer