webui: Add Firefox theme for use on Live media
The theme upstream is here: https://github.com/garrett/firefox-theme-frame

Its main aim is to shape the Firefox window to the most suitable state
on Live images:
- has a header bar for window switching
- has no tab/status/URL bars
- has no browser controls

As Firefox needs the profile folder to be writable, we need to copy the
profile template to a writable directory & use that as the profile path
for Firefox. We also need to make sure profile folders from previous
runs (possible on Live) are cleaned up.

To simplify things, we also use the same theme on non Live images for now,
with the expectation that a tweaked theme will provided later.