webui: start using payload from the bots repository
Create a second http server to serve the payload as it's in in different
path than the already served updates.img.

After creating the http servers with Popen, call p.communicate to make
sure the process is started before checking again for available ports,
in order to avoid port conflicts.