webui: fix logic for when to re-create the partitioning
Now partitioning is marked for re-creation when device selection
changes and when the user re-scans the device tree.

The latter invalidates the partitioning indirectly, by re-setting the
selected devices.

This commit also simplifies the partitioning object update in state,
as a clean state update was needed to make the above feature work as
expected.