webui: fix make rsync not updating the test VM
`make rsync` stores the new files in /usr/local/share/cockpit
whilst the original files are in /usr/share/cockpit.

So the running session doesn't recognize the new
directory in /usr/local/, that'd require a bridge restart.

This is not possible to do as that makes anaconda think that user Quit the
installation.