webui: Automatically prefill mount point for swap devices
Swap devices are not really mountable, but we need to somehow
indicate that the device will be used in the installed system and
written into /etc/fstab. For that we currenly use "swap" mount
point both in TUI mount point assignment and in Custom Spoke. For
WebUI we'll prefill the mount point for all selected devices
formatted to "swap".