webui: refactor LUKS unlocking modal after designer feedback
The new implementation makes it more clear which devices were unlocked
and which need another password.

The 'Cancel' button was also renamed to 'Close', in order to indicate
'Skip' functionality.

The password field needs to be reset between form submissions.

Tests needed to be adjusted according to the code changes.