webui: Drop global notifications in favor of the per page notifications
For the Web UI we need to keep three exception notifications types:

* Critical Error; this opens the 'Report' modal and blocks the
installation. It must be used for unexpected Exceptions from the
Backend, for example when fetching state data.

* Per page notifications; these should be created as a result of
submission of some configuration which is not accepted by the backend; the user
should be able to react to the error by changing their choices.

* Inline errors; these are feedback to the user from cheap sanity checks
for the validity of the forms.

The dropped 'onAddErrorNotification' method was adding `Toast Alerts` centrally on
the app, which does not fall under any of these categories. Replace the
two occurances of its usage with the per-page notifications.

Additionally keep the functionality of displaying per-page notifications
in the AnacondaPage component so that the code can be shared between all
pages.