packit: Build COPR for main commits
Build every commit into main into
https://copr.fedorainfracloud.org/coprs/g/rhinstaller/Anaconda-webui/

We will use that for cross-project testing (e.g. building ISOs from the
main anaconda repo containing the anaconda-webui RPM with latest merged
commit)

This now reuses the COPR project with the build after releases, but the
latter are not used anywhere to this point, so no need for a new
project.

Also stop doing copr builds on releases, these are not used and were
introduced by mistake without specific purpose in the first place.