Replace blivet-gui with cockpit-storage to be spawned on 'Modify storage'
This is the first step for the integration between cockpit-storage and
anaconda.
One basic test was introduced, to ensure that the new view is present,
but more integration tests will be introduced in a follow-up.
In follow up PR more logic will be introduced for the communication between anaconda and
cockpit, to improve user experience, by propagating mount points for
example..