flake8: Add config
We eventually want to use our standard "max line length 118" config from
other projects, but currently the tests use line lengths up to 149.
Bound it to that value at least.

Disable "module level import not at top of file" for the time
being. The tests need to be updated to Cockpit's current test shebang
for that, plus adjusting it for the local common/helpers.

Also disable "ambiguous varialbe name `l`" -- too many tests use this,
and it's a handy shortcut.