tests: unselect encryption when entering disk encryption page
The UI remembers the encryption choice from the last created
partitioning. Since the tests are non-destructive this can create test
failures, since tests expect the 'encryption' checkbox to be disabled at
start.

For fixing that this commit introduces a callback on the disk encryption
page.
For implementing the later we needed to solve a circular import
scenario, therefore we stopped importing the Installer* classes from
the Storage helpers file.