components: standardize page initialization
Pages should be able to perform initialization operations, during which
the Wizard is disabled. This is now done in each page seperately, inside
useEffect hooks calls. This commit generalizes this by extracting the
logic to a externally defined custom hook called 'usePageInit', optionally defined per-page.
For the pages not defining it, the AnacondaPage wrapper component,
is responsible to re-enable interaction in the Wizard, ie mark the page
as initialized.