Store the original deviceTree object and also the deviceTrees for the applied partitioning
Some pages show information about the actual disk state but some others
show information about the planned storage. So far we could go along
by keeping the information of one device tree, the original or the one
corresponding to the applied partitioning.

In the following commits, the need for storing the different deviceTrees becomes apparent.