tests: specify disk size also if we are using backing_file
When using backing_file at the disk creation specifying the size is not
obligatory, but if we specify it the disk image is using the specified
size.
This is quite useful for many tests scenarios, like the ones that
install in the free space on the disk another OS, as we need to control
the free space there.