Reimplement error handling using error boundaries
This is the official strategy for catching errors suggested by the
react docs [1].

This new approach also fixes an issue were crashes from asyncronous
operations or from operations at the very start of the data fetching
phase would not be caught by the error handling mechanism.

[1] https://react.dev/reference/react/Component#catching-rendering-errors-with-an-error-boundary

Note: we now don't need explicit try/catch in the whole code to catch
exceptions as these are handled by a global error boundary.

Resolves: rhbz#2308279