tests: hotfix for Anaconda Wayland migration changes
There is a new patchset to migrate Anaconda to Wayland native
application. This patchset is unfortunately quite big and complicated.
https://github.com/rhinstaller/anaconda/pull/5829

Current Web UI testing depends on updates image creation from the
Anaconda RPMs which are downloaded from Rawhide and/or COPR repos.
However, we are hitting issues for updates images in this case.

- Removal of spoke in Anaconda will not be covered by updates image.
  This is expected because RPM don't have information that some file
  was removed. Usually this is not an issue because the removed code is
  not imported. Unfortunately, here the code is imported by our
  dynamic spoke importer. That will result in traceback because
  of missing imports in the removed spoke.
- The patchset in Anaconda is adding new dependencies required to be
  able to start the Wayland environment, however, these are not covered
  when an updates image is created because we can't tell which deps are
  new and should be part of the ISO.

We should resolve this somehow by improving the workflow in the future.
For now we can continue with this hotfix.