Add interactive-defaults.ks to updates image
The interactive-defaults is must have to enable testing of web ui on the
boot.iso because we don't have support for package based installations
there. This way we are serving the tarball for the VM which will be used
as payload to overcome this missing implementation.

Do not SSH the interactive-defaults.ks into the VM when it's running but
instead add the interactive-defaults.ks to the updates.img.
The SSH copy of interactive-defaults.ks works but not reliably. The
issue is that we are copying the interactive-defaults.ks to the running
installation and if the file will land too late it won't be taken by the
installer.

To avoid this issue add the interactive-defaults.ks content to the
updates image before the VM is started.