Fix tests executed in Toolbx with host libvirt
Issue with Toolbx is that it shares the user libvirt socket so if
libvirt is started on the host system first the container will use
this host libvirt socket and won't start its own.
That works fine until container starts creating libvirt resources which
are not shared with the host. In that case libvirt fails to get these
resources.

To avoid this issue use /run/host/var/tmp when available so the /var/tmp
is still used but still accessible by the host system. The /run/host is
a way how toolbx can access host system resources so what is there is
available in both container and host system.

Similar fix is already part of the Cockpit tests:
https://github.com/cockpit-project/bots/pull/6941