Do not run firefox when remote installation is requested
Do not start firefox in the VM for remote installations.
When the user requested remote installation switch firefox instance in the VM with a plain sleep.
This currently helps to avoid the firefox crash. [1]

Also so far remote installations are used only for testing. And boot.iso
is not delivered, so the change is of no user impact.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2320964