storage: restrict access to the storage editor to the scenario selection step
Accessing the Cockpit storage editor outside the scenario selection step can
cause unexpected issues during installation.  To prevent this, the dropdown
item for the storage editor is now disabled when not on the storage entry page.

Utilize aria-disabled instead of disabled attribute, so that we can
still interact with the dropdown, and show the tooltip.