Standardize wizard page selector names to be prefix with 'anaconda-screen'
Do this because now we are exposing the selector IDs for use by
anaconda.conf, for hidden_webui_pages option.

Since this is part of an API now, we need to be consistent with the
naming of the page IDs.

We chose to have the 'anaconda-screen' prefix instead of plain 'accounts',
'review' etc for the page identifiers, as this will make it easier to find
all places where it is used & possibly make it less likely to be broken by mistake.