Fix scripts/makeupdates root detection
It currently only works if you have run certain `make` targets -
anaconda-webui.spec is not part of the version-controlled repo,
it's generated. It seems safer to look for a file under version
control.

I'm not sure the script is much use for anaconda-webui anyway -
I don't think there's a mechanism to apply changes to the jsx
files, for e.g. - but it doesn't hurt to improve this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>