tests: fix nightly action for testing latest compose
We want to test the compose untouched, that means anaconda-webui package
that is inside the compose. That means we also need to checkout the
anaconda-webui repository, before running the tests to the tag used in
the compose.
This is problematic, because our support for testing composes did not
yet land in rawhide, therefore no tag contains that currently.

Let's test against anaconda-webui from main branch, temporarily.