tests: fix old compose wiki reporting
When testing composes we checkout the git repository to the tag that
anaconda-webui has in that compose. We do that so that the tests match
the code tested.
However, there is some infra logic, that got more recently added, and is
there for reporting to fedora wiki.

Let's copy the relevant files from master till these changes are part of
the composes we are interested to test.