tests: drop testErrorHanding and adjust testBackendTranslations
For triggering an expected backend error we were using devices with same
uuids. This is now allowed in blivet, so we need to figure out another
approahc.

For testing backend translations, I converted the test to check a translated expected
error message when attempting to create partitioning.
Unfortunately the translation does not seem to work [1], so I disabled
the test for now.

For testing unhandled backend exceptions, which end up in the 'Error
report' dialog, I was not able to find another situaion, as we handle
gracefully all known errors and allow the user to take action. Therefore
the test was removed.

[1] https://issues.redhat.com/browse/INSTALLER-4142