storage: when checking available scenarios through cockpit-storage don't consider reclaim space
When the user would not create the required mount points in
cockpit-storage, they would usually get 'Free space requirements are
met' because the logic would fallback to 'Use free space' scenario,
with the reclaim space addition. This is a unwelcome behavior when existing
cockpit-storage, users are expected to create mount points there or
leave the empty space needed by the installer.