tests: add support for reporting tests on either staging or production wiki
Introduce this knowledge in the "compose" scenario. It can now be:
TEST_SCENARIO=compose-<compose_id> or
TEST_SCENARIO=compose-<compose_id>-staging

Also make test/wiki-report script configurable on which wiki instance to
report to.

Lastly introduce a parameter in the Workflow to be able to choose the
wiki instance there as well.