test: adjust windows test to use /dev/vdb
There is some problematic dependency between the following two tests:
TestStorageReclaimExistingSystemFedora.testDeletePartition
TestStorageReclaimExistingSystemWindows.testBasic

When they are run against the same VM in the above order, the Windows
test fails to detect correctly the free space.
There is some race condition in the backend, causing the Storage module
to not refresh even after re-scanning.

I am not able to resolve this so let's use a differently named disk for
the Windows test, to prevent confusing the Storage module.