cockpit-storage-integration: do not perform any API calls before new storage is in store
When the user performs operations in Cockpit Storage, we first need to re-scan the storage and
perform related tasks, such as resetting selected disks (e.g. for MD RAID).
During this transition period, API calls may attempt to access devices that are not yet valid.
Previously, we worked around this by ignoring related errors like UnknownDeviceError.
However, with the new code reorganization, we now ensure that no API calls are made until the
updated storage state is available in the store, making this workaround unnecessary.