tests: users: wait for password to be focused before starting typing
This commit c66f51f2cf38380a1008a644b3ea7313e6247abd introduces a
autofocus logic in the code for the Root Password field, after the user
enables root account creation. This was causing some flakiness in the
tests which type super fact, as it would regain focus on the first
password field, when the test would be already filling the second
password field.

Fix this by waiting for the focus before starting typing.