storage: detect locked LUKS on selected disks including whole-disk LUKS
The disks were not detected as LUKS encrypted because
getLockedLUKSDevices was checking only child devices, not disks.

Also add test scenario for mount-point-mapping unlock flow for /home on
encrypted second disk (whole disk).

Finally fix the comments on the getDeviceAncestors & getDeviceChildren to
reflect that they are not returning the device itself in the returned
list. This was probably an older behavior.

Resolves: rhbz#2396646