components/localization: Only offer ASCII-capable keyboard layouts
As discussed in the bug report, it is dangerous to allow the
user to configure a single, non-ASCII capable keyboard layout.
This makes it almost impossible to set a usable encryption
passphrase, or create usernames and passwords that use ASCII
characters (which is common practice to avoid bugs in things that
don't expect non-ASCII characters there).

The correct fix would be for webUI to handle switched layout
configurations correctly, but as a quick minimal fix, we can
simply not offer layouts that do not support ASCII input. Note
this requires an anaconda with
https://github.com/rhinstaller/anaconda/pull/6692 in it.

Resolves: rhbz#2402430

Signed-off-by: Adam Williamson <awilliam@redhat.com>