storage: encryption: show warning when the compositor layout differs from target system default
When configuring passwords show to the user the VirtualConsoleKeymap
configured value, not the CompositorSelectedLayout as the first will be
used at LUKS screen. In the disk encryption screen, display a warning
alert if the two layouts differ to inform users that the layout used for
typing the encryption passphrase during boot does not match the one
currently active in the compositor.

In order to get to this state now, the user needs to go to desktop
keyboard settings and manually change / add keyboards, after selecting
their prefered keyboard from the installer's localization screen.

However, we soon will be supporting multi-keyboard selection for non
gnome desktop variants, and this problematic state can be reached also
by changing the currently selected compositor keyboard with a shortcut.

Relevant to: rhbz#2402430