localization: keyboard: do not allow proceeding before vconsole is set
Do that check only for non Gnome environment, for Gnome keyboard
configuation is set after installation starts.

Also do not show alert for missing keyboards. On KDE these are
pre-loaded anyway, there is not way that the user ends up with no
keyboard selection.
This alert is just flickering before defaults are loaded.