components: keyboards: stop setting vconsole manually
Use GetMissingKeyboardConfigurationTask to read what will be used the
backend when installation starts.

For non Gnome, we also make sure to get sensible defaults according to
the language selection by calling SetXKeyboardDefaults.
For Gnome we are unable to set the compositor values with localed [1]
so we are not setting there, all keyboard changes need to happen by the
user through the Gnome keyboard settings.

Resolves: rhbz#2404635