Embed cockpit-networkmanager in modal dialog accessible from each page
The module is embedded in an iframe, similar as Cockpit Storage.
The entry point is the menu in the page header.

This is a MVP available only on boot.iso (non-Live image) which is not
a released product yet. It will be a subject for UX design review.
Potential targets are Fedora-ELN and Fedora Server variants.

There is a follow-up item for better Checkpoints integration
to be addressed: https://issues.redhat.com/browse/COCKPIT-1417

Resolves: INSTALLER-4524